fix: update quantity fix
